Working build. Fixed Auth Flow
